+++ src/pcloud/pcloudfs.py
@@ -142,7 +142,8 @@ class PCloudFS(FS):
         self.check()
         result = self.pcloud.createfolder(path=path)
         if result["result"] == 2004:
-            raise errors.DirectoryExists('Directory "{0}" already exists')
+            raise errors.DirectoryExists(
+                'Directory "{0}" already exists'.format(path))
         elif result["result"] != 0:
             raise errors.CreateFailed(
                 'Create of directory "{0}" failed with "{1}"'.format(